+++ .gitignore
@@ -8,3 +8,9 @@ lesson3/.idea/
 lesson4/engine/.idea/
 
 lesson4/.idea/
+
+lesson5/engine/.idea/
+
+lesson5/.idea/
+
+lesson5/engine/vendor/
