--- lesson4/engine/views/menu.php
@@ -1,4 +1,0 @@
-<a href="/">Главная</a>
-<a href="/?c=product&a=catalog">Каталог</a>
-<a href="/?c=basket">Корзина</a>
-<br>
