+++ lesson5/engine/TwigTemplates/basket.twig
@@ -1,0 +1,23 @@
+{% if basket|length == 0 %}
+    <h2>Корзина пуста</h2>
+{% else %}
+
+<table style="max-width: 768px; margin: 0 auto">
+    <tr>
+        <th style="padding: 5px">Наименование</th>
+        <th style="padding: 5px; text-align: center">Цена</th>
+        <th style="padding: 5px; text-align: center">Количество</th>
+        <th style="padding: 5px; text-align: center">Сумма</th>
+        <th></th>
+    </tr>
+    {% for item in basket %}
+        <tr>
+            <td style="padding: 5px">{{ item.name }}</td>
+            <td style="padding: 5px; text-align: center">{{ item.price }}</td>
+            <td style="padding: 5px; text-align: center">{{ item.qty }}</td>
+            <td style="padding: 5px; text-align: center">{{ item.price * item.qty }}</td>
+            <td><button>X</button></td>
+        </tr>
+    {% endfor %}
+</table>
+{% endif %}
