+++ lesson5/engine/TwigTemplates/card.twig
@@ -1,0 +1,7 @@
+<h2>Карточка</h2>
+
+
+<h3>{{ product.name }}</h3>
+<p>{{ product.price }} руб.</p>
+<p> {{ product.description }}</p>
+<button>Купить</button>
