+++ lesson5/engine/TwigTemplates/catalog.twig
@@ -1,0 +1,11 @@
+<h2>Каталог</h2>
+
+{% for item in catalog %}
+    <h3><a href="/product/card?id={{ item.id }}">{{ item.name }}</a></h3>
+    <p>{{ item.description }}</p>
+    <p>{{ item.price }}</p>
+    <button>Купить</button>
+    <hr>
+{% endfor %}
+
+<a href="/product/catalog?page={{ page }}">Далее</a>
