+++ lesson5/engine/composer.json
@@ -1,0 +1,5 @@
+{
+    "require": {
+        "twig/twig": "^3.1"
+    }
+}
