+++ lesson5/engine/engine/Render.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace app\engine;
+
+
+class Render
+{
+
+}
