+++ lesson5/engine/interfaces/IRenderer.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace app\interfaces;
+
+
+interface IRenderer
+{
+
+}
