+++ lesson5/engine/views/index.php
@@ -1,0 +1,1 @@
+Добро пожаловать!
