+++ lesson5/engine/views/menu.php
@@ -1,0 +1,4 @@
+<a href="/">Главная</a>
+<a href="/product/catalog">Каталог</a>
+<a href="/basket">Корзина</a>
+<br>
